Add visual monitor and remove the last vertion ui-test
