--- ui-tests/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "shoov-webdrivercss": "~0.0.5"
-  }
-}
