+++ visual-monitor/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "shoov-webdrivercss": "~0.1.8"
+  }
+}
