+++ visual-monitor/test/tests.js
@@ -1,0 +1,94 @@
+'use strict';
+
+var shoovWebdrivercss = require('shoov-webdrivercss');
+
+// This can be executed by passing the environment argument like this:
+// PROVIDER_PREFIX=browserstack SELECTED_CAPS=chrome mocha
+// PROVIDER_PREFIX=browserstack SELECTED_CAPS=ie11 mocha
+// PROVIDER_PREFIX=browserstack SELECTED_CAPS=iphone5 mocha
+
+var capsConfig = {
+  'chrome': {
+    'browser' : 'Chrome',
+    'browser_version' : '42.0',
+    'os' : 'OS X',
+    'os_version' : 'Yosemite',
+    'resolution' : '1024x768'
+  },
+  'ie11': {
+    'browser' : 'IE',
+    'browser_version' : '11.0',
+    'os' : 'Windows',
+    'os_version' : '7',
+    'resolution' : '1024x768'
+  },
+  'iphone5': {
+    'browser' : 'Chrome',
+    'browser_version' : '42.0',
+    'os' : 'OS X',
+    'os_version' : 'Yosemite',
+    'chromeOptions': {
+      'mobileEmulation': {
+        'deviceName': 'Apple iPhone 5'
+      }
+    }
+  }
+};
+
+var selectedCaps = process.env.SELECTED_CAPS || undefined;
+var caps = selectedCaps ? capsConfig[selectedCaps] : undefined;
+
+var providerPrefix = process.env.PROVIDER_PREFIX ? process.env.PROVIDER_PREFIX + '-' : '';
+var testName = selectedCaps ? providerPrefix + selectedCaps : providerPrefix + 'default';
+
+var baseUrl = process.env.BASE_URL ? process.env.BASE_URL : 'http://www.gizra.com';
+
+var resultsCallback = process.env.DEBUG ? console.log : shoovWebdrivercss.processResults;
+
+describe('Visual monitor testing', function() {
+
+  this.timeout(99999999);
+  var client = {};
+
+  before(function(done){
+    client = shoovWebdrivercss.before(done, caps);
+  });
+
+  after(function(done) {
+    shoovWebdrivercss.after(done);
+  });
+
+  it('should show the home page',function(done) {
+    client
+      .url(baseUrl)
+      .webdrivercss(testName + '.homepage', {
+        name: '1',
+        exclude: [],
+        remove: [],
+        hide: [],
+        screenWidth: selectedCaps == 'chrome' ? [320, 1200] : undefined,
+      }, resultsCallback)
+      .call(done);
+  });
+
+  it('should show a blog post',function(done) {
+    client
+      .url(baseUrl + '/content/shoov-monitor-twitter/')
+      .webdrivercss(testName + '.blog-post', {
+        name: '1',
+        exclude:
+          [
+            // Logo
+            '.logo'
+          ],
+        remove:
+          [
+            // Disqus plugin.
+            '#disqus_thread',
+          ],
+        screenWidth: selectedCaps == 'chrome' ? [320, 1200] : undefined,
+      }, resultsCallback)
+      .call(done);
+  });
+
+});
